server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.27.2
